codepipeline-cdk-lambda-api/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.3.0
cattrs==1.0.0
constructs==10.0.41
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
jsii==1.52.1
packaging==21.3
pkg-resources==0.0.0
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
zipp==3.6.0
